lazyGrab/requirements.txt
selenium
python-slugify
